requirements.txt
torch == 2.2.2
torchvision==0.17.2
numpy==1.26.4
matplotlib==3.8.4
Pillow==10.3.0